+++ src/components/home/HomeComp.jsx
@@ -4,13 +4,14 @@ import { sortLikePosts, showPosts } from "../../redux/modules/postWrite";
 import { useEffect } from "react";
 import { collection, getDocs, query } from "@firebase/firestore";
 import { db } from "../../../src/firebase";
+import { showMembers, sortLikeMembers } from "../../redux/modules/logReducer";
 
 const HomeComp = () => {
   const dispatch = useDispatch();
 
   useEffect(() => {
     const newArr = [];
-    const fetchData = async () => {
+    const fetchPostsData = async () => {
       const q = query(collection(db, "posts"));
       const querySnapshot = await getDocs(q);
       querySnapshot.forEach((doc) => {
@@ -18,13 +19,34 @@ const HomeComp = () => {
       });
       dispatch(showPosts(newArr));
       dispatch(sortLikePosts());
-      return newArr;
     };
-    fetchData();
+    fetchPostsData();
+
+    const fetchMemberData = async () => {
+      // q = 요청 객체
+      const q = query(collection(db, "starList"));
+      const querySnapshot = await getDocs(q);
+      const initialStarList = [];
+      querySnapshot.forEach((doc) => {
+        const data = {
+          id: doc.id,
+          ...doc.data(),
+        };
+        initialStarList.push(data);
+      });
+      dispatch(showMembers(initialStarList));
+      dispatch(sortLikeMembers());
+    };
+    fetchMemberData();
   }, []);
 
   const popPosts = useSelector((state) => state.posts);
-  console.log(popPosts);
+  console.log("popPosts", popPosts);
+
+  const StarList = useSelector((state) => {
+    return state.logReducer.members;
+  });
+  console.log("StarList", StarList);
 
   return (
     <>
